:construction: Robust unique function

+++ tests/test_unique.py
@@ -27,20 +27,20 @@ def test_unique_1d_counts(device: str):
 
 
 def test_unique_1d_inverse(device: str):
-    data_np = np.array([20, 10, 2, 3, 1, 3, 10, 20], dtype=np.int32)
+    data_np = np.array([20, 10, 2, 3, 1, 3, 10, 20], dtype=np.int64)
     unique_np, inverse_np = np.unique(data_np, return_inverse=True)
 
-    data_wp = wp.array(data_np, dtype=wp.int32, device=device)
+    data_wp = wp.array(data_np, dtype=wp.int64, device=device)
     unique_wp, inverse_wp = tw.unique.unique_1d(data_wp, return_inverse=True)
     assert np.array_equal(unique_wp.numpy(), unique_np)
     assert np.array_equal(inverse_wp.numpy(), inverse_np)
 
 
 def test_unique_1d_inverse_counts(device: str):
-    data_np = np.array([20, 10, 20, 3, 1, 3, 10, 20], dtype=np.int32)
+    data_np = np.array([20, 10, 20, 3, 1, 3, 10, 20], dtype=np.uint64)
     unique_np, inverse_np, counts_np = np.unique(data_np, return_inverse=True, return_counts=True)
 
-    data_wp = wp.array(data_np, dtype=wp.int32, device=device)
+    data_wp = wp.array(data_np, dtype=wp.uint64, device=device)
     unique_wp, inverse_wp, counts_wp = tw.unique.unique_1d(data_wp, return_inverse=True, return_counts=True)
     assert np.array_equal(unique_wp.numpy(), unique_np)
     assert np.array_equal(inverse_wp.numpy(), inverse_np)
